// sdk/packages/simplex/src/proto/mpcvault/platform/v1/api.ts
export enum ECDSAHashFunction {
  ECDSA_HASH_FUNCTION_UNSPECIFIED = 0,
  ECDSA_HASH_FUNCTION_USE_MESSAGE_DIRECTLY = 1,
  ECDSA_HASH_FUNCTION_SHA256 = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface ExecuteSigningRequestsRequest {
  /**
   * required:true
   * symbol_value:"12789f5ee5fb-07f5-470c-a2ff-081e2d6d107a"
   * comment:uuid is the UUID of the signing request. Currently only signing requests of API wallets are supported.
   */
  uuid: string;
}

export interface ExecuteSigningRequestsResponse {
  error:
    | Error
    | undefined;
  /**
   * symbol_value:"3Eq21vXNB5s86c62bVuUfTerseghjkuUqRPBmGRJhyTA"
   * comment:tx_hash is the hash of the signed transaction, only set if status is STATUS_SUCCEEDED.This does not mean that the transaction it self is successful. It only means that the signing request has been signed.
   */
  txHash: string;
  /** signatures is the signatures of the raw message, only set if status is STATUS_SUCCEEDED. */
  signatures:
    | SignatureContainer
    | undefined;
  /** signed_transaction is the signed transaction for broadcast, only set if status is STATUS_SUCCEEDED. */
  signedTransaction: string;
}
// ...
export interface SignatureContainer_ECDSASignature {
  R: string;
  S: string;
  V: string;
}
// ...
export enum EVMMessage_Type {
  TYPE_UNSPECIFIED = 0,
  TYPE_PERSONAL_SIGN = 1,
  TYPE_SIGN_TYPED_DATA = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface AptosMessage {
  /**
   * symbol_value:"0x71C7656EC7ab88bd32g47dgdhhssq245f6d8976F"
   * comment:from is the sender wallet address. It has to be in this vault.
   */
  from: string;
  /**
   * required:true
   * symbol_value:"0
   * comment:type is the type of the message.
   */
  type: AptosMessage_Type;
  content: AptosMessage_Message | undefined;
}
// ...
export interface CreateSigningRequestRequest {
  evmSendNative?: EVMSendNative | undefined;
  evmSendErc20?: EVMSendERC20 | undefined;
  evmSendCustom?: EVMSendCustom | undefined;
  btcSendNative?: BTCSendNative | undefined;
  tronSendNative?: TronSendNative | undefined;
  tronSendTrc10?: TronSendTRC10 | undefined;
  tronSendTrc20?: TronSendTRC20 | undefined;
  aptosSendNative?: AptosSendNative | undefined;
  aptosSendCoin?: AptosSendCoin | undefined;
  aptosSendCustom?: AptosSendCustom | undefined;
  suiSendNative?: SuiSendNative | undefined;
  suiSendCoin?: SuiSendCoin | undefined;
  suiSendCustom?: SuiSendCustom | undefined;
  solanaSendNative?: SolanaSendNative | undefined;
  solanaSendSplToken?: SolanaSendSPLToken | undefined;
  evmMessage?: EVMMessage | undefined;
  aptosMessage?: AptosMessage | undefined;
  solanaMessage?: SolanaMessage | undefined;
  rawMessage?: RawMessage | undefined;
  suiMessage?: SuiMessage | undefined;
  tonSendNative?: TonSendNative | undefined;
  tonSendToken?: TonSendToken | undefined;
  tonSendCustom?: TonSendCustom | undefined;
  tonMessage?: TonMessage | undefined;
  solanaSendCustom?: SolanaSendCustom | undefined;
  sonicSvmSendNative?: SolanaSendNative | undefined;
  sonicSvmSendSplToken?: SolanaSendSPLToken | undefined;
  sonicSvmSendCustom?: SolanaSendCustom | undefined;
  stellarSendNative?: StellarSendNative | undefined;
  stellarSendAsset?: StellarSendAsset | undefined;
  stellarSendCustom?: StellarSendCustom | undefined;
  stellarChangeTrust?:
    | StellarChangeTrust
    | undefined;
  /**
   * symbol_value:"send 1 eth to 0x71C7656EC7ab88bd32g47dgdhhssq245f6d8976F"
   * comment:notes is the transaction notes for the signing request.
   */
  notes:
    | string
    | undefined;
  /**
   * required:true
   * symbol_value:"9f5ee5fb-07f5-470c-a2ff-081e2d6d107a"
   * comment:vault_uuid is the UUID of the vault that the signing request will be created in.
   */
  vaultUuid:
    | string
    | undefined;
  /**
   * symbol_value:"ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIL3OzFAKNGHsMtBpTemiAI3V1AxRsdeghjcmkidagdk test"
   * comment:It is used to identify the where we should send the callback to. If empty, we will show the signing request in the MPCVault mobile app for manual signing.Note that if the signing request is created by an api wallet, this field must be set as api wallets can only be signed by the api client signer at the moment.
   */
  callbackClientSignerPublicKey:
    | string
    | undefined;
  /** comment:If true, the tx will be automatically broadcasted after signing. Not supported for Message types, RawMessage, or txs on non-system (custom EVM) chains. All other standard txs on system chains are supported. */
  broadcastTx: boolean | undefined;
}

export interface CreateSigningRequestResponse {
  /** signing_request is the signing request that was created. */
  signingRequest: SigningRequest | undefined;
  error: Error | undefined;
}
// ...
export const ExecuteSigningRequestsRequest: MessageFns<ExecuteSigningRequestsRequest> = {
  encode(message: ExecuteSigningRequestsRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.uuid !== "") {
      writer.uint32(10).string(message.uuid);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ExecuteSigningRequestsRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseExecuteSigningRequestsRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.uuid = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ExecuteSigningRequestsRequest {
    return { uuid: isSet(object.uuid) ? globalThis.String(object.uuid) : "" };
  },

  toJSON(message: ExecuteSigningRequestsRequest): unknown {
    const obj: any = {};
    if (message.uuid !== "") {
      obj.uuid = message.uuid;
    }
    return obj;
  },

  create(base?: DeepPartial<ExecuteSigningRequestsRequest>): ExecuteSigningRequestsRequest {
    return ExecuteSigningRequestsRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ExecuteSigningRequestsRequest>): ExecuteSigningRequestsRequest {
    const message = createBaseExecuteSigningRequestsRequest();
    message.uuid = object.uuid ?? "";
    return message;
  },
};
// ...
export const SignatureContainer_ECDSASignature: MessageFns<SignatureContainer_ECDSASignature> = {
  encode(message: SignatureContainer_ECDSASignature, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.R !== "") {
      writer.uint32(10).string(message.R);
    }
    if (message.S !== "") {
      writer.uint32(18).string(message.S);
    }
    if (message.V !== "") {
      writer.uint32(26).string(message.V);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SignatureContainer_ECDSASignature {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSignatureContainer_ECDSASignature();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.R = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.S = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.V = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SignatureContainer_ECDSASignature {
    return {
      R: isSet(object.R) ? globalThis.String(object.R) : "",
      S: isSet(object.S) ? globalThis.String(object.S) : "",
      V: isSet(object.V) ? globalThis.String(object.V) : "",
    };
  },

  toJSON(message: SignatureContainer_ECDSASignature): unknown {
    const obj: any = {};
    if (message.R !== "") {
      obj.R = message.R;
    }
    if (message.S !== "") {
      obj.S = message.S;
    }
    if (message.V !== "") {
      obj.V = message.V;
    }
    return obj;
  },

  create(base?: DeepPartial<SignatureContainer_ECDSASignature>): SignatureContainer_ECDSASignature {
    return SignatureContainer_ECDSASignature.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<SignatureContainer_ECDSASignature>): SignatureContainer_ECDSASignature {
    const message = createBaseSignatureContainer_ECDSASignature();
    message.R = object.R ?? "";
    message.S = object.S ?? "";
    message.V = object.V ?? "";
    return message;
  },
};
// ...
export const CreateSigningRequestResponse: MessageFns<CreateSigningRequestResponse> = {
  encode(message: CreateSigningRequestResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.signingRequest !== undefined) {
      SigningRequest.encode(message.signingRequest, writer.uint32(10).fork()).join();
    }
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateSigningRequestResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateSigningRequestResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.signingRequest = SigningRequest.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateSigningRequestResponse {
    return {
      signingRequest: isSet(object.signingRequest)
        ? SigningRequest.fromJSON(object.signingRequest)
        : isSet(object.signing_request)
        ? SigningRequest.fromJSON(object.signing_request)
        : undefined,
      error: isSet(object.error) ? Error.fromJSON(object.error) : undefined,
    };
  },

  toJSON(message: CreateSigningRequestResponse): unknown {
    const obj: any = {};
    if (message.signingRequest !== undefined) {
      obj.signingRequest = SigningRequest.toJSON(message.signingRequest);
    }
    if (message.error !== undefined) {
      obj.error = Error.toJSON(message.error);
    }
    return obj;
  },

  create(base?: DeepPartial<CreateSigningRequestResponse>): CreateSigningRequestResponse {
    return CreateSigningRequestResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<CreateSigningRequestResponse>): CreateSigningRequestResponse {
    const message = createBaseCreateSigningRequestResponse();
    message.signingRequest = (object.signingRequest !== undefined && object.signingRequest !== null)
      ? SigningRequest.fromPartial(object.signingRequest)
      : undefined;
    message.error = (object.error !== undefined && object.error !== null) ? Error.fromPartial(object.error) : undefined;
    return message;
  },
};
// ...
export type PlatformAPIService = typeof PlatformAPIService;
// ...
export const PlatformAPIClient = makeGenericClientConstructor(
  PlatformAPIService,
  "mpcvault.platform.v1.PlatformAPI",
) as unknown as {
  new (address: string, credentials: ChannelCredentials, options?: Partial<ClientOptions>): PlatformAPIClient;
  service: typeof PlatformAPIService;
  serviceName: string;
};